moved gameCreation out of GameControl
Now all that stuff is in GameFactory. The class will take care of creating fully functional controls to be launched.
Not runnable at the moment.
