+++ src/control/CreateWorld.java
@@ -1,15 +1,6 @@
 package control;
 
-import model.Character;
-import model.Item;
-import model.Obstacle;
-import model.Place;
-import view.Input;
-import view.Output;
-import model.Passage;
-import model.GameObject;
 import control.XmlParser;
-
 
 /**
  * Take information from XmlParser and use it to create several objects and relations between.
@@ -31,8 +22,6 @@ public class CreateWorld {
     control.runGame();
 
   }
-
-
